# LabRestApi/src/main/resources/docs/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "API для Лабораторной информационной системы",
    "version": "2.1"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
...
  "paths": {
    "/api/v1/tests/{id}": {
      "get": {
        "tags": [
          "Tests"
        ],
        "summary": "Получить лаб. исследования по ID",
        "description": "Возвращает лаб. исследования по указанному ID",
        "operationId": "getTestById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Лаб. исследования успешно найдены",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/TestResponseDTO"
                }
              }
            }
          },
          "404": {
            "description": "Лаб исследования не найдены",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Tests"
        ],
        "summary": "Обновить лаб. исследование",
        "description": "Обновляет данные лаб. исследования по указанному ID",
        "operationId": "updateTest",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TestRequestDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "лаб. исследование успешно обновлено",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/TestResponseDTO"
                }
              }
            }
          },
          "404": {
            "description": "лаб. исследование не найдено",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "400": {
            "description": "Некорректные данные лаб. исследования",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/tests/{id}/result": {
      "put": {
        "tags": [
          "Tests"
        ],
        "summary": "Добавляет результат лаб. исследования",
        "description": "Добавляет результат лаб. исследования по указанному ID",
        "operationId": "updateTestResult",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "string",
                "example": {
                  "result": "New result"
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Результат лаб. исследования успешно добавлен",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/TestResponseDTO"
                }
              }
            }
          },
          "404": {
            "description": "Лаб. исследование не найдено",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "400": {
            "description": "Некорректные данные для обновления лаб. исследования",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/tests": {
      "get": {
        "tags": [
          "Tests"
        ],
        "summary": "Получить все лаб. исследования",
        "description": "Возвращает список всех лаб. исследований. Реализована пагинация: по умолчанию page=0 & size=50",
        "operationId": "getAllTests",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "Zero-based page index (0..N)",
            "required": false,
            "schema": {
              "minimum": 0,
              "type": "integer",
              "default": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "description": "The size of the page to be returned",
            "required": false,
            "schema": {
              "minimum": 1,
              "type": "integer",
              "default": 50
            }
          },
          {
            "name": "sort",
            "in": "query",
            "description": "Sorting criteria in the format: property,(asc|desc). Default sort order is ascending. Multiple sort criteria are supported.",
            "required": false,
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Список тестов успешно получен",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/TestResponseDTO"
                }
              }
            }
          },
          "404": {
            "description": "Лаб. исследования не найдены",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Tests"
        ],
        "summary": "Создать новое лаб. исследование",
        "description": "Создает новое лаб. исследование на основе переданных данных",
        "operationId": "createTest",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TestRequestDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Лаб. исследование успешно создано",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/TestResponseDTO"
                }
              }
            }
          },
          "400": {
            "description": "Некорректные данные лаб. исследования",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "TestRequestDTO": {
        "required": [
          "orderId",
          "referenceValues",
          "result",
          "status",
          "testTypeId"
        ],
        "type": "object",
        "properties": {
          "orderId": {
            "maximum": 1000000,
            "minimum": 1,
            "type": "integer",
            "description": "ID заявки",
            "format": "int64"
          },
          "testTypeId": {
            "maximum": 1000000,
            "minimum": 1,
            "type": "integer",
            "description": "ID типа исследования",
            "format": "int64"
          },
          "result": {
            "maxLength": 500,
            "minLength": 0,
            "type": "string",
            "description": "Результат лабораторного исследования",
            "example": "Результат положительный"
          },
          "referenceValues": {
            "maxLength": 500,
            "minLength": 0,
            "type": "string",
            "description": "Референсные значения для лабораторного исследования",
            "example": "Значение"
          },
          "status": {
            "type": "string",
            "description": "Статус лабораторного исследования",
            "example": "PENDING",
            "enum": [
              "PENDING",
              "COMPLETED"
            ]
          }
        },
        "description": "Запрос DTO для создания и обновления лабораторного исследования"
      },
      "TestResponseDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "ID лабораторного исследования",
            "format": "int64"
          },
          "orderId": {
            "type": "integer",
            "description": "ID заявки",
            "format": "int64"
          },
          "testTypeId": {
            "type": "integer",
            "description": "ID типа исследования",
            "format": "int64"
          },
          "executionDate": {
            "type": "string",
            "description": "Дата выполнения лабораторного исследования (yyyy-MM-dd HH:mm:ss)",
            "format": "date-time"
          },
          "result": {
            "type": "string",
            "description": "Результат лабораторного исследования",
            "example": "Результат положительный"
          },
          "referenceValues": {
            "type": "string",
            "description": "Референсные значения для лабораторного исследования",
            "example": "Значение"
          },
          "status": {
            "type": "string",
            "description": "Статус лабораторного исследования",
            "example": "PENDING",
            "enum": [
              "PENDING",
              "COMPLETED"
            ]
          }
        },
        "description": "Ответ DTO для просмотра лабораторных исследований"
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "description": "Сообщение об ошибке",
            "example": "Пациентов с id-404 не найдено"
          },
          "statusCode": {
            "type": "integer",
            "description": "Код ошибки",
            "format": "int32",
            "example": 404
          }
        }
      },
...
    }
  }
}